# packages/api-spec/openapi.yaml
openapi: 3.0.3
info:
  title: LOOPIE API
  version: 0.0.1
  description: >
    Unified messaging + advertising platform. Contract mirrors docs/00-unified-data-model.md:
    Campaign (paid) and Message (owned) are separate acquisition mechanisms that both feed one
    Contact -> Lead -> Sale -> Interaction spine via a source_type/source_id pointer.
servers:
  - url: http://localhost:3001
    description: Local dev
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    Cursor:
      in: query
      name: cursor
      schema: { type: string }
      description: Opaque cursor returned by the previous page.
    Limit:
      in: query
      name: limit
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
    AffiliateId:
      in: path
      name: affiliateId
      required: true
      schema: { type: string }
    AffiliateClassId:
      in: path
      name: classId
      required: true
      schema: { type: string }
    AffiliateDealId:
      in: path
      name: dealId
      required: true
      schema: { type: string }
...
    SaleId:
      in: path
      name: saleId
      required: true
      schema: { type: string }
...
  schemas:
    PaginatedMeta:
      type: object
      required: [hasMore, nextCursor]
      properties:
        hasMore: { type: boolean }
        nextCursor: { type: string, nullable: true }
...
    AffiliateClass:
      type: object
      required: [id, businessId, name, maxAffiliateRateBps, maxManagerShareBps, createdAt]
      properties:
        id: { type: string }
        businessId: { type: string }
        name: { type: string }
        maxAffiliateRateBps: { type: integer }
        maxManagerShareBps: { type: integer }
        defaultDealId: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
    CreateAffiliateClassInput:
      type: object
      required: [name, maxAffiliateRateBps, maxManagerShareBps]
      properties:
        name: { type: string, minLength: 1, maxLength: 80 }
        maxAffiliateRateBps: { type: integer, minimum: 0, maximum: 10000 }
        maxManagerShareBps: { type: integer, minimum: 0, maximum: 10000 }
    UpdateAffiliateClassInput:
      type: object
      properties:
        name: { type: string }
        maxAffiliateRateBps: { type: integer, minimum: 0, maximum: 10000 }
        maxManagerShareBps: { type: integer, minimum: 0, maximum: 10000 }
        defaultDealId: { type: string, nullable: true }
    AffiliateDeal:
      type: object
      required: [id, businessId, name, commissionRuleType, managerShareBps, payoutCadence, isActive, createdAt]
      properties:
        id: { type: string }
        businessId: { type: string }
        classId: { type: string, nullable: true }
        name: { type: string }
        commissionRuleType: { type: string, enum: [PERCENTAGE, FIXED] }
        affiliateRateBps: { type: integer, nullable: true }
        fixedAmountMinor: { type: integer, nullable: true }
        managerShareBps: { type: integer, description: Percent of gross commission paid to the manager. Does not increase total cost. }
        eligibilityWindowDays:
          type: integer
          nullable: true
          description: Cookie-window bound measured from the referral click, not Lead creation. null = unlimited.
        payoutThresholdMinor: { type: integer, nullable: true }
        payoutCadence: { type: string, enum: [MANUAL, WEEKLY, MONTHLY] }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
    CreateAffiliateDealInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 80 }
        classId: { type: string }
        commissionRuleType: { type: string, enum: [PERCENTAGE, FIXED], description: Defaults to PERCENTAGE. }
        affiliateRateBps: { type: integer, minimum: 1, maximum: 10000 }
        fixedAmountMinor: { type: integer, minimum: 1 }
        managerShareBps: { type: integer, minimum: 0, maximum: 10000 }
        eligibilityWindowDays: { type: integer, minimum: 1 }
        payoutThresholdMinor: { type: integer, minimum: 1 }
        payoutCadence: { type: string, enum: [MANUAL, WEEKLY, MONTHLY] }
    UpdateAffiliateDealInput:
      type: object
      properties:
        name: { type: string }
        classId: { type: string, nullable: true }
        commissionRuleType: { type: string, enum: [PERCENTAGE, FIXED] }
        affiliateRateBps: { type: integer, minimum: 1, maximum: 10000 }
        fixedAmountMinor: { type: integer, minimum: 1 }
        managerShareBps: { type: integer, minimum: 0, maximum: 10000 }
        eligibilityWindowDays: { type: integer, minimum: 1, nullable: true }
        payoutThresholdMinor: { type: integer, minimum: 1, nullable: true }
        payoutCadence: { type: string, enum: [MANUAL, WEEKLY, MONTHLY] }
        isActive: { type: boolean }
    Affiliate:
      type: object
      required: [id, businessId, name, referralCode, referralUrl, classId, commissionRuleType, payoutCadence, isActive, createdAt]
      properties:
        id: { type: string }
        businessId: { type: string }
        name: { type: string }
        email: { type: string, nullable: true }
        referralCode: { type: string, description: "Globally unique, like LandingPage.slug." }
        referralUrl: { type: string, description: "GET this URL to track a click and redirect the visitor — GET /r/affiliate/{affiliateId}." }
        classId: { type: string }
        dealId: { type: string, nullable: true }
        managerId: { type: string, nullable: true }
        userId: { type: string, nullable: true }
        affiliateRateOverrideBps: { type: integer, nullable: true }
        managerShareOverrideBps: { type: integer, nullable: true }
        commissionRuleType:
          type: string
          enum: [PERCENTAGE, FIXED]
          description: Resolved live from assigned Deal (display only — historical earnings use frozen commissions).
        commissionRateBps: { type: integer, nullable: true }
        commissionFixedAmountMinor: { type: integer, nullable: true }
        managerShareBps: { type: integer }
        affiliateNetBps: { type: integer, nullable: true }
        eligibilityWindowDays: { type: integer, nullable: true }
        payoutThresholdMinor: { type: integer, nullable: true }
        payoutCadence: { type: string, enum: [MANUAL, WEEKLY, MONTHLY] }
        lastPayoutAt: { type: string, format: date-time, nullable: true }
        downlineCount: { type: integer }
        initialPassword: { type: string, description: Returned once when createLogin is true. }
        destinationLandingPageId: { type: string, nullable: true }
        destinationUrl: { type: string, nullable: true, description: Fallback redirect if no landing page is set. }
        isActive: { type: boolean }
        pausedAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
    CreateAffiliateInput:
      type: object
      required: [name, classId]
      properties:
        name: { type: string, minLength: 1, maxLength: 150 }
        email: { type: string, format: email }
        referralCode: { type: string, minLength: 1, maxLength: 40, pattern: '^[a-zA-Z0-9-]+$', description: Auto-generated if omitted. }
        classId: { type: string }
        dealId: { type: string }
        managerId: { type: string }
        createLogin: { type: boolean }
        destinationLandingPageId: { type: string }
        destinationUrl: { type: string }
    UpdateAffiliateInput:
      type: object
      properties:
        name: { type: string }
        email: { type: string, format: email, nullable: true }
        classId: { type: string }
        dealId: { type: string, nullable: true }
        managerId: { type: string, nullable: true }
        affiliateRateOverrideBps: { type: integer, minimum: 0, maximum: 10000, nullable: true }
        managerShareOverrideBps: { type: integer, minimum: 0, maximum: 10000, nullable: true }
        destinationLandingPageId: { type: string, nullable: true }
        destinationUrl: { type: string, nullable: true }
    AffiliateEarnings:
      type: object
      required: [pendingMinor, payableMinor, paidMinor, commissions, payouts]
      properties:
        pendingMinor: { type: integer }
        payableMinor: { type: integer }
        paidMinor: { type: integer }
        commissions:
          type: array
          items:
            type: object
            required: [id, amountMinor, currency, status, createdAt]
            properties:
              id: { type: string }
              amountMinor: { type: integer }
              currency: { type: string }
              status: { type: string, enum: [PENDING, PAYABLE, PAID, CANCELLED, REVERSED] }
              sourceRef: { type: string, nullable: true }
              createdAt: { type: string, format: date-time }
        payouts:
          type: array
          items:
            type: object
            required: [id, amountMinor, currency, status, createdAt]
            properties:
              id: { type: string }
              amountMinor: { type: integer }
              currency: { type: string }
              status: { type: string }
              createdAt: { type: string, format: date-time }

    # ---------- Campaign / Deployment ----------
...
    Sale:
      type: object
      required: [id, businessId, contactId, amount, date, sourceType, createdAt]
      properties:
        id: { type: string }
        businessId: { type: string }
        contactId: { type: string }
        leadId: { type: string, nullable: true }
        amount: { type: number }
        date: { type: string, format: date-time }
        productOrService: { type: string, nullable: true }
        sourceType: { type: string, enum: [MESSAGE, DEPLOYMENT, AD_UNIT, MANUAL, IMPORT] }
        sourceMessageId: { type: string, nullable: true }
        sourceDeploymentId: { type: string, nullable: true }
        sourceAdUnitId: { type: string, nullable: true }
        notes: { type: string, nullable: true }
        reversedAt:
          type: string
          format: date-time
          nullable: true
          description: Set by POST /sales/{saleId}/reverse. Not a status enum — matches Lead.closedAt/Automation.pausedAt.
        createdAt: { type: string, format: date-time }
    ReverseSaleInput:
      type: object
      properties:
        reason: { type: string }
    CreateSaleInput:
      type: object
      required: [contactId, amount, date]
      properties:
        contactId: { type: string }
        leadId: { type: string }
        amount: { type: number, minimum: 0 }
        date: { type: string, format: date-time }
        productOrService: { type: string }
        notes: { type: string }
...
paths:
...
  /affiliate-classes:
    get:
      operationId: listAffiliateClasses
      summary: List affiliate classes
      tags: [affiliates]
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Paginated classes
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/AffiliateClass' }
                  meta: { $ref: '#/components/schemas/PaginatedMeta' }
    post:
      operationId: createAffiliateClass
      summary: Create affiliate class
      tags: [affiliates]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateAffiliateClassInput' }
      responses:
        '201':
          description: Class created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/AffiliateClass' }
  /affiliate-classes/{classId}:
    get:
      operationId: getAffiliateClass
      summary: Get affiliate class
      tags: [affiliates]
      parameters:
        - $ref: '#/components/parameters/AffiliateClassId'
      responses:
        '200':
          description: Class
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/AffiliateClass' }
    patch:
      operationId: updateAffiliateClass
      summary: Update affiliate class
      tags: [affiliates]
      parameters:
        - $ref: '#/components/parameters/AffiliateClassId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateAffiliateClassInput' }
      responses:
        '200':
          description: Updated class
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/AffiliateClass' }
  /affiliate-deals:
    get:
      operationId: listAffiliateDeals
      summary: List affiliate deals
      tags: [affiliates]
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Paginated deals
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/AffiliateDeal' }
                  meta: { $ref: '#/components/schemas/PaginatedMeta' }
    post:
      operationId: createAffiliateDeal
      summary: Create affiliate deal
      tags: [affiliates]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateAffiliateDealInput' }
      responses:
        '201':
          description: Deal created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/AffiliateDeal' }
  /affiliate-deals/{dealId}:
    get:
      operationId: getAffiliateDeal
      summary: Get affiliate deal
      tags: [affiliates]
      parameters:
        - $ref: '#/components/parameters/AffiliateDealId'
      responses:
        '200':
          description: Deal
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/AffiliateDeal' }
    patch:
      operationId: updateAffiliateDeal
      summary: Update affiliate deal
      tags: [affiliates]
      parameters:
        - $ref: '#/components/parameters/AffiliateDealId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateAffiliateDealInput' }
      responses:
        '200':
          description: Updated deal
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/AffiliateDeal' }
  /affiliates:
    get:
      operationId: listAffiliates
      summary: List affiliates
      tags: [affiliates]
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Paginated affiliates
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Affiliate' }
                  meta: { $ref: '#/components/schemas/PaginatedMeta' }
    post:
      operationId: createAffiliate
      summary: Create affiliate
      tags: [affiliates]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateAffiliateInput' }
      responses:
        '201':
          description: Affiliate created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Affiliate' }
...
  /affiliates/{affiliateId}:
    get:
      operationId: getAffiliate
      summary: Get affiliate
      tags: [affiliates]
      parameters:
        - $ref: '#/components/parameters/AffiliateId'
      responses:
        '200':
          description: Affiliate
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Affiliate' }
    patch:
      operationId: updateAffiliate
      summary: Update affiliate
      tags: [affiliates]
      parameters:
        - $ref: '#/components/parameters/AffiliateId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateAffiliateInput' }
      responses:
        '200':
          description: Updated affiliate
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Affiliate' }
  /affiliates/{affiliateId}/earnings:
    get:
      operationId: getAffiliateEarnings
      summary: Frozen commissions and payouts for an affiliate
      tags: [affiliates]
      parameters:
        - $ref: '#/components/parameters/AffiliateId'
      responses:
        '200':
          description: Earnings from Commission/Payout rows — never recomputed from live Deal
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/AffiliateEarnings' }
  /affiliates/{affiliateId}/pause:
    post:
      operationId: pauseAffiliate
      summary: Pause affiliate
      tags: [affiliates]
      parameters:
        - $ref: '#/components/parameters/AffiliateId'
      responses:
        '200':
          description: Affiliate paused
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Affiliate' }
  /affiliates/{affiliateId}/resume:
    post:
      operationId: resumeAffiliate
      summary: Resume affiliate
      tags: [affiliates]
      parameters:
        - $ref: '#/components/parameters/AffiliateId'
      responses:
        '200':
          description: Affiliate resumed
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Affiliate' }

  # ---------- Campaigns ----------
...
  /sales:
    get:
      operationId: listSales
      summary: List sales
      tags: [sales]
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Paginated sales
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Sale' }
                  meta: { $ref: '#/components/schemas/PaginatedMeta' }
    post:
      operationId: createSale
      summary: Create sale
      tags: [sales]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateSaleInput' }
      responses:
        '201':
          description: Sale recorded; contact lifecycle becomes CUSTOMER and linked Lead (if any) moves to WON
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Sale' }
  /sales/{saleId}:
    get:
      operationId: getSale
      summary: Get sale
      tags: [sales]
      parameters:
        - $ref: '#/components/parameters/SaleId'
      responses:
        '200':
          description: Sale
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Sale' }
  /sales/{saleId}/reverse:
    post:
      operationId: reverseSale
      summary: Reverse sale
      tags: [sales]
      description: >
        Idempotent. Sets Sale.reversedAt, reopens the linked Lead's stage if this was its only
        active sale, and reverses the linked Commission (if any) through
        FinanceService.reverseCommission — never mutated directly.
      parameters:
        - $ref: '#/components/parameters/SaleId'
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReverseSaleInput' }
      responses:
        '200':
          description: Reversed sale
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Sale' }

  # ---------- Dashboards ----------
...